get able to output logit data from voc dataset and new nn models based on vgg preproc

+++ excuters/test_all.sh
@@ -1,4 +1,7 @@
-CHECKPOINT_ROOT=/home/cideep/Work/tensorflow/checkpoints/Link-to-checkpoints/my-fine-tuned
+PYTHON_PATH=/home/cideep/Work/tensorflow/tfenv/bin
+SCRIPT=`realpath $0`
+THISPATH=`dirname $SCRIPT`
+CHECKPOINT_ROOT=/home/cideep/Work/tensorflow/checkpoints/Link-to-checkpoints
 DATASET_ROOT=/home/cideep/Work/tensorflow/datasets/Link-to-datasets
 
 function mytest()
@@ -7,13 +10,13 @@ function mytest()
 	local DATASET_NAME=$1
 	local MODEL_NAME=$2
 	
-	PYTHON_PATH=/home/cideep/Work/tensorflow/tfenv/bin
-	SCRIPT=`realpath $0`
-	THISPATH=`dirname $SCRIPT`
-
 	DATASET_DIR=${DATASET_ROOT}/${DATASET_NAME}/tfrecord
 	CHECKPOINT_DIR=${CHECKPOINT_ROOT}/${MODEL_NAME}_${DATASET_NAME}
 	echo -e "checkpoint in ${CHECKPOINT_DIR}"
+	if [ ! -d "${CHECKPOINT_DIR}" ]; then
+		echo -e "checkpoint dir does not exits, skip this config"
+		return
+	fi
 
 	EVAL_DIR=${CHECKPOINT_DIR}/result
 	rm -r ${EVAL_DIR}
@@ -32,14 +35,13 @@ function mytest()
 }
 
 
-declare -a datasets=("cifar10" "cifar100" "VOC-2012")
-declare -a models=("inception_v4" "inception_resnet_v2")
+declare -a datasets=("voc2012" "cifar10" "cifar100")
+declare -a models=("resnet_v2_50" "resnet_v2_101" "vgg_16" "vgg_19" "inception_resnet_v2" "inception_v4")
+declare -a scopes=("resnet_v2_50" "resnet_v2_101" "vgg_16" "vgg_19" "InceptionResnetV2" "InceptionV4")
 
-for d in "${datasets[@]}"
-do
-	for m in "${models[@]}"
-	do
-		mytest "$d" "$m"
+for data in "${datasets[@]}"; do
+	for i in ${!models[@]}; do
+		mytest "$data" "${models[i]}"
 	done
 done
